style: 상수 파일 분리
- constants 폴더를 생성해 selector, store, style, element 파일로 상수들을 분리
