+++ src/components/ProductAddTab.js
@@ -4,7 +4,7 @@ import {
   createProductAddFormTemplate,
   createProductTableTemplate,
 } from '../templates/productAddMenu.js';
-import { SELECTOR } from '../constants.js';
+import SELECTOR from '../constants/selector.js';
 
 class ProductAddTab {
   render() {
