+++ src/constants/selector.js
@@ -1,4 +1,4 @@
-export const SELECTOR = {
+export default {
   // 공통
   vendingMachineApp: 'app',
   tabContentContainerId: 'tab-content-container',
@@ -45,27 +45,4 @@ export const SELECTOR = {
   dataProductNameDataset: 'data-product-name',
   dataProductPriceDataset: 'data-product-price',
   dataProductQuantityDataset: 'data-product-quantity',
-};
-
-export const PRODUCT_ADD_TAB = '상품 관리';
-export const VENDING_MACHINE_MANAGE_TAB = '잔돈 충전';
-export const PRODUCT_PURCHASE_TAB = '상품 구매';
-
-export const TAB_BUTTONS = [
-  {
-    text: PRODUCT_ADD_TAB,
-    id: SELECTOR.productAddMenuId,
-  },
-  {
-    text: VENDING_MACHINE_MANAGE_TAB,
-    id: SELECTOR.vendingMachineManageMenuId,
-  },
-  {
-    text: PRODUCT_PURCHASE_TAB,
-    id: SELECTOR.productPurchaseMenuId,
-  },
-];
-
-export const LOCAL_STORAGE_KEY = {
-  currentTab: 'currentTab',
 };
