+++ src/constants/store.js
@@ -1,0 +1,3 @@
+export default {
+  currentTab: 'currentTab',
+};
