+++ src/constants/style.js
@@ -1,0 +1,3 @@
+export default {
+  tableDataStyle: 'border: 1px solid black;padding: 10px 50px;',
+};
