+++ src/templates/vendingMachineManageMenu.js
@@ -1,4 +1,5 @@
-import { VENDING_MACHINE_MANAGE_TAB, SELECTOR } from '../constants.js';
+import SELECTOR from '../constants/selector.js';
+import { VENDING_MACHINE_MANAGE_TAB } from '../constants/element.js';
 import { createTheadTableDataTemplate } from './common.js';
 
 export const createTitleTemplate = () => `
